requirements.txt
Django~=4.1.2
djangorestframework~=3.14.0
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.1
PyJWT~=2.6.0